+++ src/CReadData.hpp
@@ -16,10 +16,10 @@ private:
     std::string line;
     std::string valueStr;
     char delimiter;
-    int dataValue;
+    double dataValue;
     int lineNum;
 //    std::vector< std::vector<int> > data;
-    std::vector<int> data;
+    std::vector<double> data;
 
     void readLine(std::stringstream &lineStream);
 public:
@@ -29,7 +29,7 @@ public:
     void readAllLines();
     void printLines(int numLines);
 //    std::vector< std::vector<int> >* getData();
-    const std::vector<int>& getData() const;
+    const std::vector<double>& getData() const;
 };
 
 #endif
